requirements.txt
streamlit
elasticsearch
openai
altair
numpy
pandas
python-dotenv
